:wheelchair: Fix tabbing through hidden menu.
Prior to this commit you could tab through anchor tags in a hidden
menu on mobile viewports without opening the menu.

This was undesirable for regular users using their keyboards for navigation
as they would have to tab through all the hidden links in the meny before
they could tab to the next visible link on the page.

This commit will fix that issue.