:art: Change img srcset="" to picture and source.
img srcset="" will allow the browser to pick the right image
but this project only provides the design and images for the
desktop and mobile screen sizes.

picture and source seem more appropriate for a project that
only has desktop and mobile designs as the browser will do exactly
what it is told when using picture.

When using the img element, resizing the browser window from a
desktop size to a mobile size will break the layout as the larger
image for the desktop size is kept by Google Chrome instead of
being swapped out for the smaller image for the mobile sized
viewport.

Also updated stylesheet to fix broken selectors due to the change.